[WIP] Implement almost working version of Finder using HTMLTree
This version has a little refactor to make the IdsSeeder be initialized
per function call.
It also adds a `HTMLNode.as_tuple/2` function to transform the HTMLNode
into tree again. Useful while collecting the results.

Using the FinderPoc made only 14 tests to break, mostly because they are
not considering combinators yet, and are not considering multi selectors
as well.