Change Ids Seeder to be simplier and faster
We don't need concurrency at this step for the most cases. So it's
better to simplify and store ids at the tree representation.

It also makes easier to know the next node id and the number of nodes.
One down side is that we need to reverse the ids list everytime we want
to traverse the tree. Not a big issue :)